Modularization, abstracting the code in controllers
