+++ server/src/server.ts
@@ -1,39 +1,11 @@
 import express from 'express';
-import { getRepository } from 'typeorm';
-import Orphanages from './models/Orphanages';
-
 import './database/conection';
+
+import rotues from './routes';
 
 const app = express();
 
 app.use(express.json());
-
-app.post('/orphanages', async (req, res) => {
-  const {
-    name,
-    latitude,
-    longitude,
-    about,
-    instructions,
-    opening_hours,
-    open_on_weekends,
-  } = req.body;
-
-  const orphanagesRepository = getRepository(Orphanages);
-
-  const orphanage = orphanagesRepository.create({
-    name,
-    latitude,
-    longitude,
-    about,
-    instructions,
-    opening_hours,
-    open_on_weekends,
-  });
-
-  await orphanagesRepository.save(orphanage);
-
-  return res.json({ message: 'Hello World' });
-});
+app.use(rotues);
 
 app.listen(3333);
